Fix: moved the AchievementWindow class out of the __init__.py
