+++ src/Ankimon/pyobj/achievement_window.py
@@ -1,0 +1,142 @@
+import json
+
+from aqt import mw
+from aqt.qt import (
+    QGridLayout,
+    QLabel,
+    QPixmap,
+    Qt,
+    QVBoxLayout,
+    QWidget,
+    QPainter,
+    )
+from PyQt6.QtGui import QIcon
+from PyQt6.QtWidgets import (
+    QLabel,
+    QVBoxLayout,
+    QWidget,
+    QScrollArea, 
+    )
+from PyQt6.QtGui import QIcon, QColor
+
+from ..resources import icon_path, badges_path, badgebag_path, badges_list_path
+
+class AchievementWindow(QWidget):
+    def __init__(self):
+        super().__init__()
+        self.read_item_file()
+        self.initUI()
+
+    def initUI(self):
+        self.setWindowIcon(QIcon(str(icon_path)))
+        self.setWindowTitle("Achievements")
+        self.layout = QVBoxLayout()  # Main layout is now a QVBoxLayout
+
+        # Create the scroll area and its properties
+        self.scrollArea = QScrollArea(self)
+        self.scrollArea.setWidgetResizable(True)
+
+        # Create a widget and layout for content inside the scroll area
+        self.contentWidget = QWidget()
+        self.contentLayout = QGridLayout()  # The layout for items
+        self.contentWidget.setLayout(self.contentLayout)
+
+        # Add the content widget to the scroll area
+        self.scrollArea.setWidget(self.contentWidget)
+
+        # Add the scroll area to the main layout
+        self.layout.addWidget(self.scrollArea)
+        self.setLayout(self.layout)
+
+    def renewWidgets(self):
+        self.read_item_file()
+        # Clear the existing widgets from the layout
+        for i in reversed(range(self.contentLayout.count())):
+            widget = self.contentLayout.itemAt(i).widget()
+            if widget:
+                widget.deleteLater()
+        row, col = 0, 0
+        max_items_per_row = 4
+        if self.badge_list is None or not self.badge_list:  # Wenn None oder leer
+            empty_label = QLabel("You dont own any badges yet.")
+            self.contentLayout.addWidget(empty_label, 1, 1)
+        else:
+            for badge_num in self.badge_list:
+                item_widget = self.BadgesLabel(badge_num)
+                self.contentLayout.addWidget(item_widget, row, col)
+                col += 1
+                if col >= max_items_per_row:
+                    row += 1
+                    col = 0
+        self.resize(700, 400)
+
+    def BadgesLabel(self, badge_num):
+        badge_path = badges_path / f"{str(badge_num)}.png"
+        frame = QVBoxLayout() #itemframe
+        with open(badges_list_path, "r", encoding="utf-8") as json_file:
+            badges = json.load(json_file)
+        achievement_description = f"{(badges[str(badge_num)])}"
+        badges_name_label = QLabel(f"{achievement_description}")
+        badges_name_label.setAlignment(Qt.AlignmentFlag.AlignCenter)
+        if badge_num < 15:
+            border_width = 93  # Example width
+            border_height = 93  # Example height
+            border_color = QColor('black')
+            border_pixmap = QPixmap(border_width, border_height)
+            border_pixmap.fill(border_color)
+            desired_width = 89  # Example width
+            desired_height = 89  # Example height
+            background_color = QColor('white')
+            background_pixmap = QPixmap(desired_width, desired_height)
+            background_pixmap.fill(background_color)
+            picture_pixmap = QPixmap(str(badge_path))
+            painter = QPainter(border_pixmap)
+            painter.drawPixmap(2, 2, background_pixmap)
+            painter.drawPixmap(5,5, picture_pixmap)
+            painter.end()  # Finish drawing
+            picture_label = QLabel()
+            picture_label.setPixmap(border_pixmap)
+        else:
+            picture_pixmap = QPixmap(str(badge_path))
+            # Scale the QPixmap to fit within a maximum size while maintaining the aspect ratio
+            max_width, max_height = 100, 100  # Example maximum sizes
+            scaled_pixmap = picture_pixmap.scaled(max_width, max_height, Qt.AspectRatioMode.KeepAspectRatio, Qt.TransformationMode.SmoothTransformation)
+            picture_label = QLabel()
+            picture_label.setPixmap(scaled_pixmap)
+        picture_label.setStyleSheet("border: 2px solid #3498db; border-radius: 5px; padding: 5px;")
+        picture_label.setAlignment(Qt.AlignmentFlag.AlignCenter)
+        frame.addWidget(picture_label)
+        frame.addWidget(badges_name_label)
+        frame_widget = QWidget()
+        frame_widget.setLayout(frame)
+
+        return frame_widget
+
+    def read_item_file(self):
+        # Read the list from the JSON file
+        with open(badgebag_path, "r", encoding="utf-8") as json_file:
+            self.badge_list = json.load(json_file)
+
+    def clear_layout(self, layout):
+        while layout.count():
+            item = layout.takeAt(0)
+            widget = item.widget()
+            if widget:
+                widget.deleteLater()
+
+    def showEvent(self, event):
+        # This method is called when the window is shown or displayed
+        self.renewWidgets()
+
+    def show_window(self):
+        # Get the geometry of the main screen
+        main_screen_geometry = mw.geometry()
+        
+        # Calculate the position to center the ItemWindow on the main screen
+        x = int(main_screen_geometry.center().x() - self.width() // 2)
+        y = int(main_screen_geometry.center().y() - self.height() // 2)
+        
+        # Move the ItemWindow to the calculated position
+        self.move(x, y)
+        
+        self.show()
